pyk cli for kompiling and running proofs with rpc booster (#958)
~Blocked on: https://github.com/runtimeverification/pyk/pull/948~

This PR brings us to be "compatible" with the K frontend `kompile` and
`kprove` commands via `pyk kompile` and `pyk prove` commands. The aim is
to be able to do no-code demos using the RPC prover of proving
functionality.

- `Kompile.default_directory()` staticmethod is added for computing the
default kompiled directory to use on CLI.
- `KProve.prove_rpc` method (and others in class) are adjusted to
respect the `--type-inference-mode` argument that `kompile` also
supports.
- A single test from `regression-new` in the K frontend is ported over.
  - The test is a simple proof test.
  - A failing version of the test is added.
- File paths in `ktest.mak` and `ktest-group.mak` are adjusted for use
in this repository.
  - A failing variant of the test is added.
- CLI `prove` is adjusted to mimic the CLI of `kprove` on the options it
does support. It outputs proof failure information differently than the
`kprove` frontend, but a lot of the same info.
- CLI `kompile` is adjusted to mimic the CLI of `kompile` on the options
it does support.

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: Tamás Tóth <tothtamas28@users.noreply.github.com>