Add command `pyk run` and tests from `regression-new` (#973)
This PR adds tests from the K `regression-new` test-suite which exercise
`krun` very simply.

- The test `regression-new/withConfig` is imported.
- The `regression-new` harness is augmented:
  - Logic for handling `krun` style tests is added in.
- The distinction between `TESTDIR` and `RESULTDIR` is removed (it's not
used in very many tests in regression-new, so they can probably be
adapted).
- The very simple `KRun.krun` command is added, which just forwards the
request to the underlying `krun` executable almost directly and
processes the output a bit.
- The CLI `pyk run` is added, which makes use of `KRun.krun`.
- The pretty-printer for single-item `KSequence` is updated to still
include the `~> .K`, and a test is added to the normal integration-tests
harness, and other test output is updated accordingly.
- The expected output of `regression-new/withConfig` is updated to
include the `<generatedCounter>` cell.

---------

Co-authored-by: devops <devops@runtimeverification.com>