More regression new tests passing (#1001)
Part of: https://github.com/runtimeverification/pyk/issues/981

This works on getting more of the `regression-new` test-suite passing:

- Adds options `--warnings [all|none]` and `-w2e` to `kompile ...`
command, and threads them through to inner calls to `kompile`.
- Remove places in the test-suite where `--warnings none` was passed to
`kprove ...`, as the entire executable is different now.
- ~Remove mentions of the strategy module (as done here:
https://github.com/runtimeverification/k/pull/4115).~
- Adjusts the new include paths for some of the `ktest.mak` framework
where it was still referring to the old directory structure.
- Makes sure we can set `VERBOSITY=[--verbose|--debug]` for both
`ktest.mak` and `ktest-fail.mak`.
- Fixes an issue with `krun.nopgm` tests were not being enumerated for
testing.
- Removes some now passing tests from the skipped list.
- Updates some test expected output.

---------

Co-authored-by: devops <devops@runtimeverification.com>