Update kompile CLI and regression-new error messages (#1012)
This PR is part of
https://github.com/runtimeverification/pyk/issues/981#issuecomment-2012635223.

Specifically:

- The `pyk kompile ...` command now accepts `--no-exc-wrap` option, and
that option is added back to the `regression-new` test-suite for calls
to `kompile`.
- The `run_process` in `pyk` now is a bit more careful about handling
failed calls to subprocess that the user wants checked. Specifically,
instead of causing a stack trace to be printed, just forwards the
subprocesses errors, prints out the command that failed and teh exit
code, and exits gracefully with the same exit code.
- A bunch of tests in `regression-new/` are updated with this new
expected output.
- Many tests in `regression-new/` are now enabled as they have the
correct output now.

---------

Co-authored-by: devops <devops@runtimeverification.com>